fix: create basket model
