+++ src/pages/basketPage/basketPageController.ts
@@ -1,6 +1,7 @@
-import { getAnonymousFlow, getCookie, getPasswordFlow, postAnonymousFlow, postPasswordFlow } from "../../shared/API";
+import { getCookie } from "../../shared/API";
 import { basketButtonController } from "../../shared/router";
-import { CartResponse, CartResponseItem, lineItem } from "../../types/interfaces/basketPage";
+import { CartResponse, lineItem } from "../../types/interfaces/basketPage";
+import { addProductToCart, createCart, deleteProductFromCart, getCartData } from "./basketModel";
 
 export function checkAuthorization(): boolean {
   const user = getCookie('access_token');
@@ -8,57 +9,6 @@ export function checkAuthorization(): boolean {
     return true
   }
   return false
-}
-
-export async function createCart(): Promise<void> {
-  if (checkAuthorization()) {
-    await postPasswordFlow(`/me/carts`, {
-      "currency": "USD"
-    })
-  }
-  await postAnonymousFlow(`/carts`, {
-    "currency": "USD"
-  })
-}
-
-export async function getProductInCart(): Promise<CartResponseItem> {
-  let cartResult: CartResponse
-  if (checkAuthorization()) {
-    cartResult = await getPasswordFlow(`/me/carts`);
-    if (!cartResult.results[0]) {
-      createCart()
-      cartResult = await getPasswordFlow(`/me/carts`);
-    }
-  } else {
-    cartResult = await getAnonymousFlow(`/carts`)
-  }
-  return cartResult.results[0]
-}
-
-export async function getCartData(): Promise<CartResponse> {
-  let response: CartResponse
-  if (checkAuthorization()) {
-    response = await getPasswordFlow(`/me/carts`);
-  } else {
-    response = await getAnonymousFlow(`/carts`);
-  }
-  return response
-};
-
-export async function addProductToCart(data: object, cartId: string): Promise<void> {
-  if (checkAuthorization()) {
-    await postPasswordFlow(`/me/carts/${cartId}`, data)
-  } else {
-    await postAnonymousFlow(`/carts/${cartId}`, data)
-  }
-}
-
-export async function deleteProductFromCart(data: object, cartId: string, cartDataVersion: number): Promise<void> {
-  if (checkAuthorization()) {
-    await postPasswordFlow(`/me/carts/${cartId}?version=${cartDataVersion}`, data)
-    return
-  }
-  await postAnonymousFlow(`/carts/${cartId}?version=${cartDataVersion}`, data)
 }
 
 export async function sendDataToCart(e: Event): Promise<void> {
